Initial commit of v0.3.0
Original homepage here: http://public.dev.nuxeo.com/~ben/plush/
That version hasn't been updated since 2007 and doesn't work with current PyLucene releases.
The svn server provided still has an A record in DNS but doesn't respond on either tcp/80 or tcp/443.

Forking and hosting on github so my changes to update it are not lost.